hijack `write eeprom` to select page of EEPROM greater than 1KB
As mentioned many times, LGT8F328p emulates EEPROM using Flash memory. 
By default its size is set to 1KB, but it can be changed to 2Kb, 4KB or 8KB.
And because the ISP pretend to be connected to an ATmega328p, AVRdude will not allow to dump EEPROM greater than 1024.

To workaround this limiation, the `write eeprom 1023 0x??` command can be used to tell to the ISP which 1KB page of the EEPROM will be dumped, and what is the size of the emulated EEPROM.

Note : The EEPROM address 1023 (or 0x3ff) is the address of the last byte of the 1KB page. As mentioned previously, the 2 last bytes of each 1KB page are read-only anyway, because they are used by the EEPROM controller to store a magic number.

Example :

`write eeprom 1023 0x04` will select page 0 of a 4KB EEPROM.

`write eeprom 1023 0x38` will select page 3 of a 8KB EEPROM.

`dump eeprom 0 1024` will displays the content according to these parameters.

By default, the ISP consider the eeprom is 1KB. (which is equivalent to `write eeprom 1023 0x01`.)